Rayshaun (#5)
* set ignore filter

* Structure for repo to match project

* Practicing 'proper' project file structure
