--- Node_1-link.h
@@ -1,7 +1,0 @@
-#pragma once
-
-template <class T>
-struct Link {
-	T data;
-	Link * next;
-};
