+++ TexasHoldEm/src/main/include/Node_1-link.h
@@ -1,0 +1,7 @@
+#pragma once
+
+template <class T>
+struct Link {
+	T data;
+	Link * next;
+};
